# Speed up Cholesky Jacobi matrices