--- a/src/clenshaw.jl
+++ b/src/clenshaw.jl
@@ -306,14 +306,17 @@ end
 
 getindex(M::Clenshaw, k::Int, j::Int) = M[k:k,j][1]
 
-function getindex(S::Symmetric{T,<:Clenshaw{<:Any}}, kr::AbstractUnitRange, j::Integer) where T
-    m = max(kr.start,jr.start,kr.stop,jr.stop)
-    return getindex(parent(S),1:m,1:m)[kr,j]
+function getindex(S::Symmetric{T,<:Clenshaw{<:Any}}, k::Integer, jr::AbstractUnitRange) where T
+    m = max(jr.start,jr.stop,k)
+    return Symmetric(getindex(S.data,1:m,1:m),S.uplo)[k,jr]
 end
-
+function getindex(S::Symmetric{T,<:Clenshaw{<:Any}}, kr::AbstractUnitRange, j::Integer) where T
+    m = max(kr.start,kr.stop,j)
+    return Symmetric(getindex(S.data,1:m,1:m),S.uplo)[kr,j]
+end
 function getindex(S::Symmetric{T,<:Clenshaw{<:Any}}, kr::AbstractUnitRange, jr::AbstractUnitRange) where T
     m = max(kr.start,jr.start,kr.stop,jr.stop)
-    return getindex(S.data,1:m,1:m)[kr,jr]
+    return Symmetric(getindex(S.data,1:m,1:m),S.uplo)[kr,jr]
 end
 
 transposelayout(M::ClenshawLayout) = LazyBandedMatrices.LazyBandedLayout()