WIP to address https://github.com/JuliaApproximation/ClassicalOrthogonalPolynomials.jl/issues/167. It will also fix some docstring issues.

For now a lot of speed is gained by creating a dedicated method for getindex of `Symmetric{Clenshaw}`, `CholeskyJacobiData` as well as better resizing in `cholesky_jacobimatrix`. 

For low polynomial degree of the weight the method is approaching acceptable speeds again, though even with Symmetric Clenshaw now as fast as regular Clenshaw the Clenshaw evals are still a major bottleneck:
```julia
julia> P = Normalized(Legendre());
julia> x = axes(P,1);
julia> J = jacobimatrix(P);
julia> wf(x) = (2 .+ x.^4);
julia> W = Symmetric(P \ (wf.(x) .* P));
# raw Clenshaw timing
julia> @time W[1:1000,1:1000];
  0.027092 seconds (18 allocations: 485.281 KiB)
# Cholesky X timing
julia> Jchol = cholesky_jacobimatrix(W, P);
julia> @time Jchol[1:1000,1:1000];
  0.061909 seconds (4.57 k allocations: 8.131 MiB)
```

Of course for high polynomial degree this causes major slow downs, although there are clearly also other things to optimize:
```julia
julia> r = 0.5;
julia> lmin, lmax = (1-sqrt(r))^2,  (1+sqrt(r))^2;
julia> P = Normalized(chebyshevu(lmin..lmax));
julia> x = axes(P,1);
julia> J = jacobimatrix(P);
julia> wf(x) = 1/(x*r);
# raw Clenshaw timing
julia> W = Symmetric(P \ (wf.(x) .* P));
julia> @time W[1:1000,1:1000];
  3.562982 seconds (18 allocations: 10.764 MiB)
# Cholesky X timing
julia> Jchol = cholesky_jacobimatrix(W, P);
julia> @time Jchol[1:1000,1:1000];
  5.413114 seconds (4.60 k allocations: 19.039 MiB)
```


Todo list:
- [x] While symmetric and non-symmetric Clenshaw are now equal in speed, Clenshaw evaluation is still very slow which constitutes 50%+ of the compute time spent in `cholesky_jacobimatrix`. 
- [x] Other optimizations and reduce allocations.
- [x] Also make changes for QR method once Cholesky is satisfactory
- [x] Add tests