@TSGut 2024-01-22
Ok, like I said based on profiling and benchmarking the remaining cost is in evaluating the Clenshaw matrix and performing its Cholesky decompositions.

(I still need to add tests and make tests pass, so not ready for merge yet)

@dlfivefifty How do you want to proceed with this? Is there hope for speeding up Clenshaw or the Cholesky of a Clenshaw? If not then I think this is bottlenecked at the speed of this PR.

To be fair, the computation is fast for weight modifications of low to medium degree weights, just not high degree where Clenshaw starts to struggle. We can to a 1 million size Jacobi matrix sized block in under a second for a degree 2 weight:
```julia
julia> P = Normalized(Legendre());

julia> x = axes(P,1);

julia> J = jacobimatrix(P);

julia> wf2(x) = (2 .+ x.^2);

julia> W2 = Symmetric(P \ (wf2.(x) .* P));

julia> Jchol2 = cholesky_jacobimatrix(W2, P);

julia> @time Jchol2[1:1000000,1:1000000];
  0.762198 seconds (16.00 M allocations: 1.103 GiB, 5.34% gc time)
```

Barring Clenshaw and Cholesky speed-ups, I suspect implementing the rational case is the next best thing.