@dlfivefifty 2024-01-22
That speed looks fine